Support pre-Debian stretch versions of debootstrap
On Debian jessie and earlier versions, the `--merged-usr` argument is
unavailable.  We'll test for this by trying to use it with no other
arguments.  This fails on both versions, but on recent versions it
shows usage information first.  On earlier versions it just throws the
error about the unrecognized option.