Handle earlier versions of sfdisk
Earlier versions of sfdisk always return four partition entries for
`msdos` type partition tables even when fewer have been defined.  The
undefined ones are shown as starting on the zero sector and having
zero size.  We'll avoid trying to bind to these non-partitions by
checking for this.  Debian jessie has a version of sfdisk this
affects.