Improve listing of dependencies
We now list `parted` as this is not installed by default on Debian.
And we now differentiate between Debian versions in our sample
invocation of `apt-get`.  On Debian jessie `systemd-nspawn` is
included with `systemd`, but `dbus` is not installed and is needed to
run `systemd-nspawn`.  On Debian stretch/sid `systemd-nspawn` has been
broken off into a separate `systemd-container` package.